requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
openai==1.75.0
pillow==11.1.0
opencv-python-headless==4.10.0.84
numpy==2.0.2
websockets==15.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
matplotlib==3.10.0
cryptography==43.0.3
scikit-image==0.24.0
scipy==1.14.1
python-dotenv